requirements.txt
pandas
numpy
torch
scikit-learn
faiss-cpu
fastapi
uvicorn
joblib